feat: Moved Version class Moved version class to namespace ListenToIt.Updater.Package for global use
