+++ ListenToIt.Core/ListenToIt.Updater/ListenToIt.Updater.csproj
@@ -65,6 +65,7 @@
     <Compile Include="CmdOptions\UpdateOptions.cs" />
     <Compile Include="Install\Installer.cs" />
     <Compile Include="Package\PackageInfo.cs" />
+    <Compile Include="Package\Version.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UpdateDownloader.cs" />
