+++ ListenToIt.Core/ListenToIt.Updater/Package/Version.cs
@@ -1,0 +1,42 @@
+﻿using System;
+
+namespace ListenToIt.Updater.Package {
+    public class Version {
+        public Version(int major, int minor, int patch, int revision, VersionSuffix suffix) {
+            Major = major;
+            Minor = minor;
+            Patch = patch;
+            Revision = revision;
+            Suffix = suffix;
+        }
+
+        /// <summary>
+        /// Construct a Version instance.
+        /// </summary>
+        /// <param name="rawVersion">rawVersion should follow "Major.Minor.Patch.Revision-Suffix" format.</param>
+        public Version(string rawVersion) {
+            Major = Convert.ToInt32(rawVersion.Split('-')[0].Split('.')[0]);
+            Minor = Convert.ToInt32(rawVersion.Split('-')[0].Split('.')[1]);
+            Patch = Convert.ToInt32(rawVersion.Split('-')[0].Split('.')[2]);
+            Revision = Convert.ToInt32(rawVersion.Split('-')[0].Split('.')[3]);
+            Suffix = (VersionSuffix) Enum.Parse(typeof(VersionSuffix), rawVersion.Split('-')[0], true);
+        }
+            
+        public int Major { get; }
+        public int Minor { get; }
+        public int Patch { get; }
+        public int Revision { get; }
+        public VersionSuffix Suffix { get; }
+
+        public bool IsNewerThan(Version ver) {
+            if (Major > ver.Major) return true;
+            if (Minor > ver.Major) return true;
+            if (Patch > ver.Patch) return true;
+            return Revision > ver.Revision;
+        }
+
+        public string GetVersionString() => $"{Major}.{Minor}.{Patch}.{Revision}-{Suffix.ToString().ToLower()}";
+
+        public enum VersionSuffix { Stable, Beta }
+    }
+}
